cleanup auth
