+++ services/client-app/src/components/nav-main.tsx
@@ -1,6 +1,7 @@
 'use client'
 
-import { ChevronRight, type LucideIcon } from 'lucide-react'
+import { ChevronRight } from 'lucide-react'
+import type { LucideIcon } from 'lucide-react'
 
 import {
   Collapsible,
@@ -21,16 +22,16 @@ import {
 export function NavMain({
   items,
 }: {
-  items: {
+  items: Array<{
     title: string
     url: string
     icon?: LucideIcon
     isActive?: boolean
-    items?: {
+    items?: Array<{
       title: string
       url: string
-    }[]
-  }[]
+    }>
+  }>
 }) {
   return (
     <SidebarGroup>
