+++ services/client-app/src/routes/index.tsx
@@ -1,85 +1,62 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { useState } from 'react'
-import logo from '../logo.svg'
-import { useAuth } from '../lib/auth-context'
-import { LoginForm, SignupForm, UserProfile } from '../components/AuthForms'
+import AuthContainer from '@/components/auth/auth-container'
+import { useAuth } from '@/lib/auth-context'
+import { ScrollArea } from '@/components/ui/scroll-area'
+import { SearchInput } from '@/components/ui/search-input'
 
 export const Route = createFileRoute('/')({
-  component: App,
+  component: RouteComponent,
 })
 
-function App() {
+function RouteComponent() {
   const { currentUser } = useAuth()
-  const [showSignup, setShowSignup] = useState(false)
 
   return (
-    <div className="text-center">
-      <header className="min-h-screen flex flex-col items-center justify-center bg-[#282c34] text-white text-[calc(10px+2vmin)]">
-        <img
-          src={logo}
-          className="h-[40vmin] pointer-events-none animate-[spin_20s_linear_infinite]"
-          alt="logo"
-        />
-
-        <div className="mt-8 max-w-md w-full">
-          {currentUser ? (
-            <UserProfile />
-          ) : (
-            <>
-              {showSignup ? (
-                <>
-                  <SignupForm />
-                  <p className="mt-4">
-                    Already have an account?{' '}
-                    <button
-                      onClick={() => setShowSignup(false)}
-                      className="text-[#61dafb] hover:underline"
-                    >
-                      Log in
-                    </button>
-                  </p>
-                </>
-              ) : (
-                <>
-                  <LoginForm />
-                  <p className="mt-4">
-                    Don't have an account?{' '}
-                    <button
-                      onClick={() => setShowSignup(true)}
-                      className="text-[#61dafb] hover:underline"
-                    >
-                      Sign up
-                    </button>
-                  </p>
-                  <p className="mt-2 text-sm text-gray-500 dark:text-gray-400">
-                    Auth Domain: {import.meta.env.VITE_FIREBASE_AUTH_DOMAIN}
-                  </p>
-                </>
-              )}
-            </>
-          )}
+    <div className=" min-h-[93vh] bg-black/3 rounded-2xl ">
+      <ScrollArea className="h-[93vh] w-full rounded-2xl relative">
+        <div className="grid grid-cols-2 sm:grid-cols-3 gap-6 p-5 xl:grid-cols-6 rounded-2xl">
+          <SearchInput />
+          {/* simulation only for layout*/}
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
+          <div className="bg-black/10 h-[42vh] rounded-2xl"></div>
         </div>
+      </ScrollArea>
 
-        <div className="mt-8">
-          <a
-            className="text-[#61dafb] hover:underline"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-          <span className="mx-2">|</span>
-          <a
-            className="text-[#61dafb] hover:underline"
-            href="https://tanstack.com"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn TanStack
-          </a>
+      {currentUser ? (
+        ''
+      ) : (
+        <div className="h-screen">
+          <AuthContainer />{' '}
         </div>
-      </header>
+      )}
     </div>
   )
 }
